+++ src/LlmTornado.Agents.Samples/Common/Utility/FileIOUtility.cs
@@ -4,7 +4,7 @@ using System.Linq;
 using System.Text;
 using System.Threading.Tasks;
 
-namespace LlmTornado.Agents.Samples.Common;
+namespace LlmTornado.Agents.Samples.Common.Utility;
 
 public class FileIOUtility
 {
