+++ src/LlmTornado.Agents.Samples/DataModels/WebSearchItem.cs
@@ -1,0 +1,13 @@
+﻿namespace LlmTornado.Agents.Samples.DataModels;
+
+public struct WebSearchItem
+{
+    public string reason { get; set; }
+    public string query { get; set; }
+
+    public WebSearchItem(string reason, string query)
+    {
+        this.reason = reason;
+        this.query = query;
+    }
+}
