+++ src/LlmTornado.Agents.Samples/DataModels/WebSearchPlan.cs
@@ -1,0 +1,10 @@
+﻿namespace LlmTornado.Agents.Samples.DataModels;
+
+public struct WebSearchPlan
+{
+    public WebSearchItem[] items { get; set; }
+    public WebSearchPlan(WebSearchItem[] items)
+    {
+        this.items = items;
+    }
+}
