+++ src/LlmTornado.Agents.Samples/ResearchAgent/ResearchAgent.cs
@@ -1,72 +1,14 @@
-﻿using LlmTornado.Agents;
-using LlmTornado.Agents.ChatRuntime;
-using LlmTornado.Agents.ChatRuntime.Orchestration;
+﻿using LlmTornado.Agents.ChatRuntime.Orchestration;
 using LlmTornado.Agents.ChatRuntime.RuntimeConfigurations;
 using LlmTornado.Agents.DataModels;
+using LlmTornado.Agents.Samples.DataModels;
 using LlmTornado.Chat;
 using LlmTornado.Chat.Models;
 using LlmTornado.Code;
 using LlmTornado.Responses;
-using System;
-using System.Collections.Generic;
-using System.Linq;
-using System.Numerics;
-using System.Text;
-using System.Threading.Tasks;
 
 namespace LlmTornado.Agents.Samples.ResearchAgent;
 
-#region Data Models
-public struct WebSearchPlan
-{
-    public WebSearchItem[] items { get; set; }
-    public WebSearchPlan(WebSearchItem[] items)
-    {
-        this.items = items;
-    }
-}
-
-public struct WebSearchItem
-{
-    public string reason { get; set; }
-    public string query { get; set; }
-
-    public WebSearchItem(string reason, string query)
-    {
-        this.reason = reason;
-        this.query = query;
-    }
-}
-
-public struct ReportData
-{
-    public string ShortSummary { get; set; }
-    public string FinalReport { get; set; }
-    public string[] FollowUpQuestions { get; set; }
-    public ReportData(string shortSummary, string finalReport, string[] followUpQuestions)
-    {
-        this.ShortSummary = shortSummary;
-        this.FinalReport = finalReport;
-        this.FollowUpQuestions = followUpQuestions;
-    }
-
-    public override string ToString()
-    {
-        return $@"
-Summary: 
-{ShortSummary}
-
-Final Report: 
-{FinalReport}
-
-
-Follow Up Questions: 
-{string.Join("\n", FollowUpQuestions)}
-";
-    }
-}
-
-#endregion
 
 public class ResearchAgentConfiguration : OrchestrationRuntimeConfiguration
 {
